Minor fixes as found via Win32 building on another CI

+++ src/Lightweight/DataBinder/SqlNumeric.hpp
@@ -91,7 +91,7 @@ struct SqlNumeric
 
     [[nodiscard]] constexpr LIGHTWEIGHT_FORCE_INLINE float ToFloat() const noexcept
     {
-        return float(ToUnscaledValue()) / std::pow(10, Scale);
+        return float(ToUnscaledValue()) / std::powf(10, Scale);
     }
 
     [[nodiscard]] constexpr LIGHTWEIGHT_FORCE_INLINE double ToDouble() const noexcept
